requirements.txt
astroid==2.3.3
attrs==19.3.0
bandit==1.6.2
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
cryptography==3.3.1
Deprecated==1.2.7
entrypoints==0.3
flake8==3.7.9
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pbr==5.4.4
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyGithub==1.46
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
PyYAML==5.3.1
requests==2.23.0
schedule==0.6.0
selenium==3.141.0
six==1.14.0
smmap==3.0.1
stevedore==1.32.0
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
